requirements.txt
requests~=2.28.1
selenium==4.7.2 # 4.1.0 for touch_actions
setuptools==65.6.3
selenium_profiles
pillow~=9.4
trio